+++ app/src/main/java/com/orange/ods/app/ui/components/progress/ProgressCircular.kt
@@ -33,10 +33,7 @@ import androidx.compose.ui.res.stringResource
 import com.orange.ods.app.R
 import com.orange.ods.app.ui.components.utilities.ComponentCustomizationBottomSheetScaffold
 import com.orange.ods.app.ui.utilities.composable.CodeImplementationColumn
-import com.orange.ods.app.ui.utilities.composable.CodeParameter
 import com.orange.ods.app.ui.utilities.composable.FunctionCallCode
-import com.orange.ods.app.ui.utilities.composable.PredefinedParameter
-import com.orange.ods.app.ui.utilities.composable.StringRepresentationParameter
 import com.orange.ods.app.ui.utilities.composable.Subtitle
 import com.orange.ods.compose.component.OdsComponent
 import com.orange.ods.compose.component.chip.OdsChoiceChip
@@ -99,11 +96,9 @@ fun ProgressCircular() {
                     FunctionCallCode(
                         name = OdsComponent.OdsCircularProgressIndicator.name,
                         exhaustiveParameters = false,
-                        parameters = mutableListOf<CodeParameter>().apply {
-                            if (type.value == ProgressCustomizationState.Type.Determinate) {
-                                add(StringRepresentationParameter("progress", determinateProgressValue))
-                            }
-                            if (hasLabel) add(PredefinedParameter.Label(text))
+                        parameters = {
+                            if (type.value == ProgressCustomizationState.Type.Determinate) stringRepresentation("progress", determinateProgressValue)
+                            if (hasLabel) label(text)
                         }
                     )
                 }